requirements.txt
Django==1.11.14
djangorestframework==3.8.2
markdown==2.6.11
coreapi==2.3.3
djangorestframework-jwt==1.11.0
django-cors-headers==2.3.0
tblib==1.3.2
